shuffling test components around, adding additional state machine tests
